backend/requirements.txt
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
Pillow==11.3.0
gunicorn==23.0.0

